refactor: move utils into their own folder & split into files
