+++ src/rules/utils/__tests__/detectJestVersion.test.ts
