+++ src/rules/utils/enums.ts
@@ -1,0 +1,46 @@
+export enum ModifierName {
+  not = 'not',
+  rejects = 'rejects',
+  resolves = 'resolves',
+}
+
+export enum EqualityMatcher {
+  toBe = 'toBe',
+  toEqual = 'toEqual',
+  toStrictEqual = 'toStrictEqual',
+}
+
+export enum DescribeAlias {
+  'describe' = 'describe',
+  'fdescribe' = 'fdescribe',
+  'xdescribe' = 'xdescribe',
+}
+
+export enum TestCaseName {
+  'fit' = 'fit',
+  'it' = 'it',
+  'test' = 'test',
+  'xit' = 'xit',
+  'xtest' = 'xtest',
+}
+
+export enum HookName {
+  'beforeAll' = 'beforeAll',
+  'beforeEach' = 'beforeEach',
+  'afterAll' = 'afterAll',
+  'afterEach' = 'afterEach',
+}
+
+export enum DescribeProperty {
+  'each' = 'each',
+  'only' = 'only',
+  'skip' = 'skip',
+}
+
+export enum TestCaseProperty {
+  'each' = 'each',
+  'concurrent' = 'concurrent',
+  'only' = 'only',
+  'skip' = 'skip',
+  'todo' = 'todo',
+}
