+++ src/rules/utils/getNodeName.ts
@@ -1,0 +1,41 @@
+import {
+  AST_NODE_TYPES,
+  TSESTree,
+} from '@typescript-eslint/experimental-utils';
+import { DescribeAlias, TestCaseName } from './enums';
+import { getAccessorValue, isSupportedAccessor } from './misc';
+import {
+  JestFunctionCallExpression,
+  JestFunctionIdentifier,
+  JestFunctionMemberExpression,
+  JestFunctionName,
+} from './types';
+
+const joinNames = (a: string | null, b: string | null): string | null =>
+  a && b ? `${a}.${b}` : null;
+
+export function getNodeName(
+  node:
+    | JestFunctionCallExpression<DescribeAlias | TestCaseName>
+    | JestFunctionMemberExpression<JestFunctionName>
+    | JestFunctionIdentifier<JestFunctionName>
+    | TSESTree.TaggedTemplateExpression,
+): string;
+export function getNodeName(node: TSESTree.Node): string | null;
+export function getNodeName(node: TSESTree.Node): string | null {
+  if (isSupportedAccessor(node)) {
+    return getAccessorValue(node);
+  }
+
+  switch (node.type) {
+    case AST_NODE_TYPES.TaggedTemplateExpression:
+      return getNodeName(node.tag);
+    case AST_NODE_TYPES.MemberExpression:
+      return joinNames(getNodeName(node.object), getNodeName(node.property));
+    case AST_NODE_TYPES.NewExpression:
+    case AST_NODE_TYPES.CallExpression:
+      return getNodeName(node.callee);
+  }
+
+  return null;
+}
