+++ src/rules/utils/scope.ts
@@ -1,0 +1,43 @@
+import { TSESLint } from '@typescript-eslint/experimental-utils';
+
+const collectReferences = (scope: TSESLint.Scope.Scope) => {
+  const locals = new Set();
+  const unresolved = new Set();
+
+  let currentScope: TSESLint.Scope.Scope | null = scope;
+
+  while (currentScope !== null) {
+    for (const ref of currentScope.variables) {
+      const isReferenceDefined = ref.defs.some(def => {
+        return def.type !== 'ImplicitGlobalVariable';
+      });
+
+      if (isReferenceDefined) {
+        locals.add(ref.name);
+      }
+    }
+
+    for (const ref of currentScope.through) {
+      unresolved.add(ref.identifier.name);
+    }
+
+    currentScope = currentScope.upper;
+  }
+
+  return { locals, unresolved };
+};
+
+export const scopeHasLocalReference = (
+  scope: TSESLint.Scope.Scope,
+  referenceName: string,
+) => {
+  const references = collectReferences(scope);
+
+  return (
+    // referenceName was found as a local variable or function declaration.
+    references.locals.has(referenceName) ||
+    // referenceName was not found as an unresolved reference,
+    // meaning it is likely not an implicit global reference.
+    !references.unresolved.has(referenceName)
+  );
+};
